Fixed an error with inputing the sentimentScore variable as a double in Datastore class then casting it as a float when it is to be returned. Fixed front end error 500 message so when a message is submitted, it is now displayed on the user page with the sentimentScore of the message.

+++ src/main/java/com/google/codeu/data/Datastore.java
@@ -47,8 +47,7 @@ public class Datastore {
     messageEntity.setProperty("text", message.getText());
     messageEntity.setProperty("timestamp", message.getTimestamp());
     messageEntity.setProperty("recipient", message.getRecipient());
-    messageEntity.setProperty(
-        "sentimentScore", message.getSentimentScore()); // stores the sentimentScore
+    messageEntity.setProperty("sentimentScore", message.getSentimentScore());
 
     datastore.put(messageEntity);
   }
@@ -75,9 +74,9 @@ public class Datastore {
         String user = (String) entity.getProperty("user");
         String text = (String) entity.getProperty("text");
         long timestamp = (long) entity.getProperty("timestamp");
-        float sentimentScore = (float) entity.getProperty("sentimentScore");
+        double sentimentScore = (double) entity.getProperty("sentimentScore");
 
-        Message message = new Message(id, user, text, timestamp, recipient, sentimentScore);
+        Message message = new Message(id, user, text, timestamp, recipient, (float) sentimentScore);
         messages.add(message);
       } catch (Exception e) {
         System.err.println("Error reading message.");
@@ -109,22 +108,9 @@ public class Datastore {
         String recipient = (String) entity.getProperty("recipient");
         String text = (String) entity.getProperty("text");
         long timestamp = (long) entity.getProperty("timestamp");
-        float sentimentScore =
-            (float)
-                entity.getProperty(
-                    "sentimentScore"); // datastore doesnt stupport floats so change from double to
-        // float
-        // potential issue because sentimentScore starts as a float so does it have to be placed in
-        // DS as a double?
+        double sentimentScore = (double) entity.getProperty("sentimentScore");
 
-        Message message =
-            new Message(
-                id,
-                user,
-                text,
-                timestamp,
-                recipient,
-                sentimentScore); // adds sentimentScore argument to the message to be returned
+        Message message = new Message(id, user, text, timestamp, recipient, (float) sentimentScore);
         messages.add(message);
       } catch (Exception e) {
         System.err.println("Error reading message.");
